Move Input Enums to upper level
Changes
- Moved enums KeyAxis and KeyMapping from class KeyInput to upper level
- Moved all input-related components to package mayonez.input
