+++ src/test/resources/testassets/foobar/foobar.json
@@ -1,4 +1,4 @@
 {
-    "date": "2021-09-01",
-    "time": "23:52:46.100669"
+    "date": "2021-09-02",
+    "time": "17:27:47.677565"
 }
