+++ src/test/resources/testassets/foobar/foobar.txt
@@ -1,2 +1,2 @@
-date=2021-09-01
-time=23:52:46.090513
+date=2021-09-02
+time=17:27:47.666322
